+++ NEWS.md
@@ -1,5 +1,10 @@
 # datawizard (devel)
 
+CHANGES
+
+* `recode_into()` gains an `overwrite` argument to skip overwriting already
+  recoded cases when multiple recode patterns apply to the same case.
+
 BUG FIXES
 
 * Fixed issues in `data_write()` when writing labelled data into SPSS format